add open action, refactor

+++ rnote-ui/src/appwindow/imp.rs
@@ -338,31 +338,31 @@ impl RnAppWindow {
         let obj = self.obj();
 
         if let Some(removed_id) = self.recovery_source_id.borrow_mut().replace(glib::source::timeout_add_seconds_local(self.recovery_interval_secs.get(),
-                clone!(@weak obj as appwindow => @default-return glib::source::Continue(false), move || {
+            clone!(@weak obj as appwindow => @default-return glib::source::Continue(false), move || {
                     let canvas = appwindow.active_tab().canvas();
-                    if canvas.output_file().is_some() && appwindow.autosave() {
-                        // Delete recovery files from disk to avoid suggesting the user an outdated file on next boot
-                        canvas.imp().recovery_file_metadata.borrow_mut().iter_mut().for_each(|meta| {
-                            meta.delete();
-                        });
-                        // We keep the metadata path in the canvas to make sure it doesnt change when the user toggles autosave.
-                        // This would otherwise lead to confusing timestamps on next launch.
-                        return glib::source::Continue(true);
-                    }
-
                     glib::MainContext::default().spawn_local(clone!(@weak canvas, @weak appwindow => async move {
-                        if !canvas.unsaved_changes_recovery() {return;}
-                        let tmp_file = canvas.get_or_generate_tmp_file();
-                        appwindow.overlays().start_pulsing_progressbar();
-                        canvas.set_recovery_in_progress(true);
-                        if let Err(e) = canvas.save_document_to_file(&tmp_file).await {
-                            log::error!("saving document failed, Error: `{e:?}`");
-                            appwindow.overlays().dispatch_toast_error(&gettext("Saving document failed"));
+                        if canvas.output_file().is_some() && appwindow.autosave() {
+                            // Delete recovery files from disk to avoid suggesting the user an outdated file on next boot
+                            if let Some(meta) = &*canvas.imp().recovery_metadata.borrow_mut() {
+                                meta.delete();
+                            };
+                            canvas.set_recovery_paused(true);
+                            // We keep the metadata path in the canvas to make sure it doesnt change when the user toggles autosave.
+                            // This would otherwise lead to confusing timestamps on next launch.
+                        }else{
+                            canvas.set_recovery_paused(false);
+                            if !canvas.unsaved_changes_recovery() {return;}
+                            let removery_file = canvas.get_or_generate_recovery_file();
+                            appwindow.overlays().start_pulsing_progressbar();
+                            canvas.set_recovery_in_progress(true);
+                            if let Err(e) = canvas.save_document_to_file(&removery_file).await {
+                                log::error!("saving document failed, Error: `{e:?}`");
+                                appwindow.overlays().dispatch_toast_error(&gettext("Saving document failed"));
+                            }
+                            canvas.set_recovery_in_progress(false);
+                            appwindow.overlays().finish_progressbar();
                         }
-                        canvas.set_recovery_in_progress(false);
-                        appwindow.overlays().finish_progressbar();
-                    }
-                ));
+                    }));
                 glib::source::Continue(true)
             }))) {
                 removed_id.remove();
